Simplify filesystem plugin for ACLs
We no longer have unified ACL codebase with FreeBSD and so we
no longer need to have separate acl_base and acl_linux files.
